🗺️ Atlas: Extract AtomicPayload to separate module
